# Various scheme-related optimizations.