--- a/include/ada/common_defs.h
+++ b/include/ada/common_defs.h
@@ -275,4 +275,12 @@ namespace ada {
 #define ADA_ASSERT_EQUAL(LHS, RHS, MESSAGE)
 #define ADA_ASSERT_TRUE(COND)
 #endif
+
+
+
+#ifdef ADA_VISUAL_STUDIO
+#define ADA_ASSUME(COND) __assume(COND)
+#else
+#define ADA_ASSUME(COND) do { if (!(COND)) __builtin_unreachable(); } while (0)
+#endif
 #endif // ADA_COMMON_DEFS_H