The main trick that this is using is that instead of, for example, passing the view "http" and then later creating the temporary "http:" which gets added to our buffer, we just pass straight up  "http:" and avoid the messy business entirely. Also, I created new methods to set the URL to file type, very fast (instead of passing the unnecessary string "file").

Apple M2, LLVM 14, benchdata.

Before:
```
Benchmark                                        Time             CPU   Iterations        GHz cycle/byte cycles/url instructions/byte instructions/cycle instructions/ns instructions/url     ns/url      speed  time/byte   time/url      url/s
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------

BasicBench_AdaURL_aggregator              25379323 ns     25377464 ns           28    3.40142    9.73549    845.617            42.563            4.37194         14.8708         3.69699k    248.607 342.355M/s  2.92095ns  253.711ns 3.94149M/s
BasicBench_AdaURL_aggregator_just_parse   25524323 ns     25523074 ns           27    3.32936    9.76305    848.011            42.482            4.35131         14.4871         3.68996k    254.707 340.401M/s  2.93771ns  255.167ns   3.919M/s
```

After:
```
Benchmark                                        Time             CPU   Iterations        GHz cycle/byte cycles/url instructions/byte instructions/cycle instructions/ns instructions/url     ns/url      speed  time/byte   time/url      url/s
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
BasicBench_AdaURL_aggregator              22431659 ns     22431677 ns           31    3.39537    8.73146    758.408           39.4446            4.51752         15.3387         3.42613k    223.365 387.314M/s  2.58189ns  224.261ns  4.4591M/s
BasicBench_AdaURL_aggregator_just_parse   22537055 ns     22537065 ns           31    3.40045    8.76141    761.009           39.4039            4.49744         15.2933         3.42259k    223.797 385.502M/s  2.59402ns  225.314ns 4.43824M/s
```